Move CRD interfaces to core plugin
Move the CRD interfaces to the core plugin, so that we can use them
within the plugin. This allows us to remove some duplicated code and to
improve the home page located in the core plugin.
