+++ plugins/applications/src/components/page/Application.tsx
@@ -19,9 +19,8 @@ import { QueryObserverResult, useQuery } from 'react-query';
 import React, { useState } from 'react';
 import { UsersIcon } from '@patternfly/react-icons';
 
-import { ExternalLink, Title } from '@kobsio/plugin-core';
+import { ExternalLink, IApplication, Title } from '@kobsio/plugin-core';
 import { DashboardsWrapper } from '@kobsio/plugin-dashboards';
-import { IApplication } from '../../utils/interfaces';
 
 interface IApplicationsParams {
   cluster: string;
