+++ plugins/core/src/components/app/account/AccountTeams.tsx
@@ -1,16 +1,17 @@
-import { Gallery, GalleryItem, PageSection, PageSectionVariants } from '@patternfly/react-core';
+import { Gallery, GalleryItem } from '@patternfly/react-core';
 import React from 'react';
 import { useQuery } from 'react-query';
 
-import { IAuthProfile, IAuthProfileTeam } from '../../context/AuthContext';
 import AccountTeamsItem from './AccountTeamsItem';
+import { ITeam } from '../../../crds/team';
+import { IUser } from '../../../crds/user';
 
 export interface IAccountTeamsProps {
-  user: IAuthProfile;
+  user: IUser;
 }
 
 const AccountTeams: React.FunctionComponent<IAccountTeamsProps> = ({ user }: IAccountTeamsProps) => {
-  const { isError, isLoading, data } = useQuery<IAuthProfileTeam[], Error>(['users/teams', user], async () => {
+  const { isError, isLoading, data } = useQuery<ITeam[], Error>(['users/teams', user], async () => {
     try {
       const response = await fetch(`/api/plugins/users/teams?cluster=${user.cluster}&namespace=${user.namespace}`, {
         body: JSON.stringify({
@@ -39,21 +40,19 @@ const AccountTeams: React.FunctionComponent<IAccountTeamsProps> = ({ user }: IAc
   }
 
   return (
-    <PageSection variant={PageSectionVariants.default}>
-      <Gallery hasGutter={true}>
-        {data.map((team, index) => (
-          <GalleryItem key={index}>
-            <AccountTeamsItem
-              cluster={team.cluster}
-              namespace={team.namespace}
-              name={team.name}
-              description={team.description}
-              logo={team.logo}
-            />
-          </GalleryItem>
-        ))}
-      </Gallery>
-    </PageSection>
+    <Gallery hasGutter={true}>
+      {data.map((team, index) => (
+        <GalleryItem key={index}>
+          <AccountTeamsItem
+            cluster={team.cluster}
+            namespace={team.namespace}
+            name={team.name}
+            description={team.description}
+            logo={team.logo}
+          />
+        </GalleryItem>
+      ))}
+    </Gallery>
   );
 };
