+++ plugins/core/src/components/app/plugins/Plugins.tsx
@@ -1,8 +1,8 @@
 import { Gallery, GalleryItem } from '@patternfly/react-core';
 import React, { useContext } from 'react';
 
-import { AuthContext, IAuthContext } from '../../context/AuthContext';
-import { IPluginData } from '../../context/PluginsContext';
+import { AuthContext, IAuthContext } from '../../../context/AuthContext';
+import { IPluginData } from '../../../context/PluginsContext';
 import PluginItem from './PluginItem';
 
 export interface IPluginsProps {
