+++ plugins/dashboards/src/index.ts
@@ -4,7 +4,6 @@ import './assets/dashboards.css';
 
 import icon from './assets/icon.png';
 
-import { IPlugin as IPluginInternal, IReference } from './utils/interfaces';
 import Page from './components/page/Page';
 import Panel from './components/panel/Panel';
 
@@ -19,5 +18,3 @@ const dashboardsPlugin: IPluginComponents = {
 export default dashboardsPlugin;
 
 export * from './components/dashboards/DashboardsWrapper';
-export type IDashboardReference = IReference;
-export type IPlugin = IPluginInternal;
