+++ plugins/users/src/utils/interfaces.ts
@@ -1,3 +1,5 @@
+// IPanelOptions is the interface for the options for a users panel used within a dashboard. It contains all required
+// fields to reference a team.
 export interface IPanelOptions {
   cluster?: string;
   namespace?: string;
